backend/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.1
django-allauth==0.51.0
django-bootstrap5==22.1
django-cors-headers==4.2.0
django-crispy-forms==1.14.0
django-filter==22.1
django-rest-swagger==2.2.0
django-smart-selects==1.6.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.1
openapi-codec==1.3.2
packaging==23.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
PyYAML==6.0.1
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.19.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
types-cryptography==3.3.23
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
